Split Cardano.CLI.Legacy.Options to a module per parser
